fix norm

+++ bonus/gnl_bonus/get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: mdaghouj <mdaghouj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 10:24:27 by mdaghouj          #+#    #+#             */
-/*   Updated: 2025/02/17 18:43:35 by mdaghouj         ###   ########.fr       */
+/*   Updated: 2025/02/18 15:46:03 by mdaghouj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 # endif
 
 char	*get_next_line(int fd);
-size_t	ft_strlen(const char *s);
+int		ft_strlen(const char *s);
 int		check_line(int byte_read, char *buffer);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strdup(const char *s);
